Ajust improvement code~

+++ backend/adopt.pets/src/main/java/com/br/adopt/pets/service/PetService.java
@@ -1,42 +1,82 @@
 package com.br.adopt.pets.service;
 
-import com.br.adopt.pets.model.Pet;
+import java.util.List;
+import java.util.stream.Collectors;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.br.adopt.pets.dtos.PetCompletDTO;
+import com.br.adopt.pets.dtos.PetDTO;
+import com.br.adopt.pets.exception.ResourceBadRequestException;
+import com.br.adopt.pets.exception.ResourceNotFoundException;
+import com.br.adopt.pets.model.Pet;
 import com.br.adopt.pets.repository.PetRepository;
 
-import java.util.ArrayList;
-
+import lombok.extern.slf4j.Slf4j;
+@Slf4j
 @Service
 public class PetService {
 
 	@Autowired
 	private PetRepository repository;
 
-	public Pet addNew(Pet novo){
-		if(novo.getNome() != null && novo.getFoto() != null){
-			return repository.save(novo);
+	public PetCompletDTO addNewPet(PetDTO novo){
+		
+		log.info("criado usuario Pet:{}",novo);
+		
+		return new PetCompletDTO(repository.save(new Pet(novo)));
+		
+		
+	}
+
+	public List<PetCompletDTO> getPetAll(){
+		
+		log.info("Buscando todos usuarios Pet");
+		
+		List<Pet> list =repository.findAll();
+		
+		log.info("Buscando todos usuarios Pet com sucesso");
+		
+		return list.stream().map(x ->new PetCompletDTO(x)).collect(Collectors.toList());
+		
+	}
+
+	public PetCompletDTO getPetById(Long id){
+		
+		log.info("Buscando usuario Pet pelo id :{}",id);	
+		
+		Pet pet= repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Não encontrado Pet"));
+		  
+		log.info("Buscando usuario Pet pelo id :{} com sucesso",id);
+	
+		return new PetCompletDTO(pet);
+	}
+
+	public PetDTO changePet(PetDTO dados, Long id){
+		
+		log.info("Alterando informações do adotante :{} , id adotante:{}",dados,id);
+		
+		if(id == null) {
+			throw new ResourceBadRequestException("Não informou id do Pet");
 		}
-		return null;
+		
+		Pet pet = repository.save(new Pet(dados));
+		
+		
+		log.info("Alterando com sucesso as informações do Pet :{} , id pet:{}");
+		
+			return new PetDTO(pet);
 	}
 
-	public ArrayList<Pet> getAll(){
-		return (ArrayList<Pet>) repository.findAll();
-	}
-
-	public Pet getById(Long id){
-		return (Pet) repository.findById(id).orElse(null);
-	}
-
-	public Pet changeCad(Pet dados){
-		if(dados.getId() != null && dados.getNome() != null){
-			return repository.save(dados);
+	public void deletePet(Long id){
+		
+		if(id== null) {
+			throw new ResourceBadRequestException("Não informou id do Pet ");
 		}
-		return null;
-	}
-
-	public void deleteCad(Long id){
+		
+		getPetById(id);
+		
 		repository.deleteById(id);
 	}
